+++ AppsV2/OceanDemo/Entities/ODPEntity.h
@@ -9,7 +9,7 @@
                       error:(NSError **)error
                            ;
 
-- (void) update:(const float)frameTime;
+- (void) update:(const double)frameTime;
 - (void) render;
 
 @end
